requirements.txt
torch==2.9.1
torchvision==0.24.1
diffusers==0.32.2
transformers==4.48.3
accelerate==1.3.0
streamlit==1.42.0
Pillow==11.3.0
python-docx==1.1.2
spacy==3.8.3
sentencepiece==0.2.0
protobuf==5.29.3
safetensors==0.5.2
huggingface-hub==0.28.1
